FINERACT-2169: Refactor currencies api resource;
