+++ fineract-provider/src/main/java/org/apache/fineract/organisation/monetary/api/CurrenciesApiResourceSwagger.java
@@ -19,8 +19,6 @@
 package org.apache.fineract.organisation.monetary.api;
 
 import io.swagger.v3.oas.annotations.media.Schema;
-import java.util.Collection;
-import org.apache.fineract.organisation.monetary.data.CurrencyData;
 
 /**
  * Created by sanyam on 14/8/17.
@@ -49,30 +47,6 @@ public final class CurrenciesApiResourceSwagger {
         public String nameCode;
         @Schema(example = "US Dollar ($)")
         public String displayLabel;
-    }
-
-    @Schema(description = "GetCurrenciesResponse")
-    public static final class GetCurrenciesResponse {
-
-        private GetCurrenciesResponse() {
-
-        }
-
-        public Collection<CurrencyData> selectedCurrencyOptions;
-        public Collection<CurrencyData> currencyOptions;
-    }
-
-    @Schema(description = "PutCurrenciesRequest")
-    public static final class PutCurrenciesRequest {
-
-        private PutCurrenciesRequest() {
-
-        }
-
-        @Schema(example = "[\"KES\",\n" + "        \"BND\",\n" + "        \"LBP\",\n" + "        \"GHC\",\n" + "        \"USD\",\n"
-                + "        \"XOF\",\n" + "        \"AED\",\n" + "        \"AMD\"]")
-        public String[] currencies;
-
     }
 
     @Schema(description = "PutCurrenciesResponse")
